+++ js/BigPicture/BigPicture.js
@@ -1,59 +1,10 @@
 import {isEscapeKey} from '../Utils/Utils.js';
+import {comments} from '../Comments/Comments.js';
 
 const bigPicture = document.querySelector('.big-picture');
 const body = document.body;
 
-const createComment = (obj) => {
-  const comment = document.createElement('li');
-  const avatar = document.createElement('img');
-  const socialText = document.createElement('p');
-  comment.classList.add('social__comment');
-  avatar.classList.add('social__picture');
-  avatar.alt = obj.name;
-  avatar.width = 35;
-  avatar.height = 35;
-  avatar.src = obj.avatar;
-  socialText.classList.add('social__text');
-  socialText.textContent = obj.message;
-
-  comment.appendChild(avatar);
-  comment.appendChild(socialText);
-
-  return comment;
-};
-
-const viewSocialComments = (arr) => {
-  const socialComments = document.querySelector('.social__comments');
-  const fragment = document.createDocumentFragment();
-
-  socialComments.textContent = '';
-  arr.forEach((item) => {
-    fragment.appendChild(createComment(item));
-  });
-  socialComments.appendChild(fragment);
-
-  return socialComments;
-};
-
-const closeBigPicture = () => {
-  bigPicture.classList.add('hidden');
-  body.classList.remove('modal-open');
-  document.removeEventListener('keydown', onPopupEscKeydown);
-};
-
-const onPopupEscKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closeBigPicture();
-  }
-};
-
-const onPopupCloseBigPicture = (evt) => {
-  evt.preventDefault();
-  closeBigPicture();
-};
-
-const showBigPicture = (obj) => {
+export const showBigPicture = (obj) => {
   bigPicture.classList.remove('hidden');
   body.classList.add('modal-open');
 
@@ -66,11 +17,20 @@ const showBigPicture = (obj) => {
   bigPicture.querySelector('.social__comment-count').classList.add('hidden');
   bigPicture.querySelector('.comments-loader').classList.add('hidden');
 
-  viewSocialComments(obj.comments);
-  document.addEventListener('keydown', onPopupEscKeydown);
-  cancelBigPicture.addEventListener('click', onPopupCloseBigPicture);
+  comments(obj.comments);
+  document.addEventListener('keydown', closeBigPictureEscKeydown);
+  cancelBigPicture.addEventListener('click', closeBigPicture);
 };
 
-export const onClickOpenBigPicturePopup = (obj) => {
-  showBigPicture(obj);
+const closeBigPicture = (evt) => {
+  evt.preventDefault();
+  bigPicture.classList.add('hidden');
+  body.classList.remove('modal-open');
+  document.removeEventListener('keydown', closeBigPictureEscKeydown);
+};
+
+const closeBigPictureEscKeydown = (evt) => {
+  if (isEscapeKey(evt)) {
+    closeBigPicture(evt);
+  }
 };
