+++ js/main.js
@@ -1,9 +1,8 @@
 import {initialData} from './Mock/InitialData.js';
 import {createData} from './Mock/CreateData.js';
-import {fillGallery} from './DrawPreview/DrawPreview.js';
+import {getGallery} from './Gallery/Gallery.js';
 
 import './UploadFile/UploadFile.js';
 import '../nouislider/nouislider.js';
 
-const data = createData(initialData);
-fillGallery(data);
+getGallery(createData(initialData));
